rename : entities/coverLetter/ui/common-style => entities/qnaboard/ui
