+++ client/src/entities/qnaboard/ui/common-style.ts
